+++ packages/core/android/src/main/java/io/sentry/react/RNSentryModuleImpl.java
@@ -304,7 +304,7 @@ public class RNSentryModuleImpl {
     }
 
     private SentryReplayOptions getReplayOptions(@NotNull ReadableMap rnOptions) {
-        @NotNull final SentryReplayOptions androidReplayOptions = new SentryReplayOptions();
+        @NotNull final SentryReplayOptions androidReplayOptions = new SentryReplayOptions(false);
 
         @Nullable final ReadableMap rnExperimentsOptions = rnOptions.getMap("_experiments");
         if (rnExperimentsOptions == null) {
@@ -317,7 +317,7 @@ public class RNSentryModuleImpl {
 
         androidReplayOptions.setSessionSampleRate(rnExperimentsOptions.hasKey("replaysSessionSampleRate")
                 ? rnExperimentsOptions.getDouble("replaysSessionSampleRate") : null);
-        androidReplayOptions.setErrorSampleRate(rnExperimentsOptions.hasKey("replaysOnErrorSampleRate")
+        androidReplayOptions.setOnErrorSampleRate(rnExperimentsOptions.hasKey("replaysOnErrorSampleRate")
                 ? rnExperimentsOptions.getDouble("replaysOnErrorSampleRate") : null);
 
         if (!rnOptions.hasKey("mobileReplayOptions")) {
@@ -328,12 +328,12 @@ public class RNSentryModuleImpl {
             return androidReplayOptions;
         }
 
-        androidReplayOptions.setRedactAllText(!rnMobileReplayOptions.hasKey("maskAllText") || rnMobileReplayOptions.getBoolean("maskAllText"));
-        androidReplayOptions.setRedactAllImages(!rnMobileReplayOptions.hasKey("maskAllImages") || rnMobileReplayOptions.getBoolean("maskAllImages"));
+        androidReplayOptions.setMaskAllText(!rnMobileReplayOptions.hasKey("maskAllText") || rnMobileReplayOptions.getBoolean("maskAllText"));
+        androidReplayOptions.setMaskAllImages(!rnMobileReplayOptions.hasKey("maskAllImages") || rnMobileReplayOptions.getBoolean("maskAllImages"));
 
         final boolean redactVectors = !rnMobileReplayOptions.hasKey("maskAllVectors") || rnMobileReplayOptions.getBoolean("maskAllVectors");
         if (redactVectors) {
-            androidReplayOptions.addClassToRedact("com.horcrux.svg.SvgView"); // react-native-svg
+            androidReplayOptions.addMaskViewClass("com.horcrux.svg.SvgView"); // react-native-svg
         }
 
         return androidReplayOptions;
@@ -380,9 +380,17 @@ public class RNSentryModuleImpl {
     }
 
     public void fetchNativeAppStart(Promise promise) {
-        final Map<String, Object> measurement = InternalSentrySdk.getAppStartMeasurement();
+        fetchNativeAppStart(promise, InternalSentrySdk.getAppStartMeasurement(), logger, AppStartMetrics.getInstance().isAppLaunchedInForeground());
+    }
 
-        WritableMap mutableMeasurement = (WritableMap) RNSentryMapConverter.convertToWritable(measurement);
+    protected void fetchNativeAppStart(Promise promise, final Map<String, Object> appStartMeasurement, ILogger logger, boolean isAppLaunchedInForeground) {
+        if (!isAppLaunchedInForeground) {
+            logger.log(SentryLevel.WARNING, "Invalid app start data: app not launched in foreground.");
+            promise.resolve(null);
+            return;
+        }
+
+        WritableMap mutableMeasurement = (WritableMap) RNSentryMapConverter.convertToWritable(appStartMeasurement);
         mutableMeasurement.putBoolean("has_fetched", hasFetchedAppStart);
 
         // This is always set to true, as we would only allow an app start fetch to only